feature: Add download_library.py
Add script that with the help of the spotipy library it can to download all the Spotify metadata of songs in the library of the given user. It will first ask the user to authorize the interaction with its library, after which it requires the redirect URL so it can continue to download 50 songs at a time until it can't find any more of them or until 10,000 songs are downloaded. The results are saved into the path './resources/user_library.json'.